Entrega final

+++ src/main/java/com/disney/preaceleracion/service/PersonajeService.java
@@ -2,7 +2,13 @@ package com.disney.preaceleracion.service;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
+
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+import javax.persistence.criteria.CriteriaBuilder;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Predicate;
+import javax.persistence.criteria.Root;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,23 +24,83 @@ public class PersonajeService {
 	@Autowired
 	PersonajeRepository repository;
 	
-	public List<PersonajeDto> buscarPersonaje(Map<String,String> query){
-		List<PersonajeDto> resp = new ArrayList<>();
-		return resp;
+	@PersistenceContext
+	EntityManager em;
+	
+	public List<PersonajeDto> listarPersonajes() {
+		List<Personaje> personajeListEntity = new ArrayList<>();
+		repository.findAll().forEach(personajeListEntity::add);
+		List<PersonajeDto> personajesListDto = new ArrayList<>();
+		personajeListEntity.stream().map(c -> personajesListDto.add(PersonajeDto.getEntity(c))).toArray();
+		return personajesListDto;
 	}
 	
-	public Boolean crearPersonje(PersonajeDto personajeDto) {
-		Personaje personaje = personajeDto.buildEntidad();
-		repository.saveAndFlush(personaje);
-		return true;
+	public String crearPersonje(PersonajeDto personajeDto) {
+		if(verificarDatosPersonaje(personajeDto)) {
+			Personaje personaje = personajeDto.buildEntidad();
+			repository.saveAndFlush(personaje);
+			return "Pelicula Agregada con exito";
+		}else {
+			return "Error: Datos Incompletos";
+		}
 	}
+
 	public Boolean actualizarPersonaje(PersonajeDto personajeDto, int id) {
-		Personaje personaje = personajeDto.buildEntidadId(id);
-		repository.saveAndFlush(personaje);
-		return true;
+		if (repository.getById(id) != null) {
+			Personaje personaje = personajeDto.buildEntidadId(id);
+			repository.saveAndFlush(personaje);
+			return true;
+		}else {
+			return false;
+		}
 	}
 	public Boolean borrarPersonaje(int id) {
-		repository.deleteById(id);
-		return true;
+		if (repository.getById(id) != null) {
+			repository.deleteById(id);
+			return true;
+		}else {
+			return false;
+		}
 	}
+
+	public List<PersonajeDto> listarPersonajesPorParams(String nombre, String edad, String idmovie) {
+		CriteriaBuilder qb = em.getCriteriaBuilder();
+		CriteriaQuery<Personaje> cq = qb.createQuery(Personaje.class);
+	    Root<Personaje> customer = cq.from(Personaje.class);
+
+	    //Constructing list of parameters
+	    List<Predicate> predicates = new ArrayList<Predicate>();
+
+	    //Adding predicates in case of parameter not being null
+	    if (nombre != null) {
+	        predicates.add(
+	                qb.equal(customer.get("nombre"), nombre));
+	    }
+	    if (edad != null) {
+	        predicates.add(
+	                qb.equal(customer.get("edad"), edad));
+	    }
+	    if (idmovie != null) {
+	        predicates.add(
+	                qb.equal(customer.get("pelicula_id"), idmovie));
+	    }
+	    cq.select(customer)
+	            .where(predicates.toArray(new Predicate[]{}));
+	    
+	    List<Personaje> personajeListEntity = new ArrayList<>();
+	    personajeListEntity = em.createQuery(cq).getResultList();
+		List<PersonajeDto> personajesListDto = new ArrayList<>();
+		personajeListEntity.stream().map(c -> personajesListDto.add(PersonajeDto.getEntity(c))).toArray();
+		return personajesListDto;
+	}
+	
+
+	private boolean verificarDatosPersonaje(PersonajeDto personajeDto) {
+		if (personajeDto.getImagen() != null && personajeDto.getNombre() != null){
+			return true;			
+		}else {
+			return false;
+		}
+	}
+	
 }
